Carregamento dos dados na Aba de Games

+++ app/src/main/java/br/senai/sp/jandira/testefragment/model/Game.kt
@@ -7,6 +7,6 @@ data class Game(
     var gameName: String,
     var gameImage: Drawable,
     var gameDescription: String,
-    var gameRating: Double,
+    var gameRating: Float,
     var consoles: List<Console>
 )
